refact: moved implementation related classes to separate package
