+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/dynapi/data/implementation/DynapiInMemoryOntModel.java
@@ -1,4 +1,4 @@
-package edu.cornell.mannlib.vitro.webapp.dynapi.data.types;
+package edu.cornell.mannlib.vitro.webapp.dynapi.data.implementation;
 
 import java.io.StringReader;
