+++ api/src/test/java/edu/cornell/mannlib/vitro/webapp/dynapi/RESTEndpointIntegrationTest.java
@@ -33,7 +33,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import edu.cornell.mannlib.vitro.webapp.application.ApplicationUtils;
-import edu.cornell.mannlib.vitro.webapp.dynapi.data.types.DynapiModelFactory;
+import edu.cornell.mannlib.vitro.webapp.dynapi.data.implementation.DynapiModelFactory;
 
 import org.apache.http.HttpHeaders;
 import org.apache.http.entity.ContentType;
