+++ api/src/test/java/edu/cornell/mannlib/vitro/webapp/dynapi/components/ModelWriterTest.java
@@ -15,13 +15,14 @@ import edu.cornell.mannlib.vitro.webapp.dynapi.data.Data;
 import edu.cornell.mannlib.vitro.webapp.dynapi.data.DataStore;
 import edu.cornell.mannlib.vitro.webapp.dynapi.data.TestView;
 import edu.cornell.mannlib.vitro.webapp.dynapi.data.conversion.InitializationException;
+import edu.cornell.mannlib.vitro.webapp.dynapi.data.implementation.DynapiInMemoryOntModel;
 
 public class ModelWriterTest {
 
 	private static final String RETRACTION = "retraction";
 	private static final String TARGET = "target";
 	private static final String ADDITION = "addition";
-	public static final String MODEL_CONVERSION_CLASS = "edu.cornell.mannlib.vitro.webapp.dynapi.data.types.DynapiInMemoryOntModel";
+	public static final String MODEL_CONVERSION_CLASS = DynapiInMemoryOntModel.class.getCanonicalName();
 	private OntModelImpl targetModel;
 	private Parameter addition;
 	private Parameter target;
