+++ api/src/test/java/edu/cornell/mannlib/vitro/webapp/dynapi/data/conversion/ParameterConverterTest.java
@@ -8,6 +8,7 @@ import java.util.Arrays;
 
 import org.junit.Test;
 
+import edu.cornell.mannlib.vitro.webapp.dynapi.data.implementation.DynapiArrayList;
 import edu.cornell.mannlib.vitro.webapp.dynapi.data.types.ArrayParameterType;
 import edu.cornell.mannlib.vitro.webapp.dynapi.data.types.ImplementationConfig;
 import edu.cornell.mannlib.vitro.webapp.dynapi.data.types.ImplementationType;
@@ -80,7 +81,7 @@ public class ParameterConverterTest {
 
 	@Test
 	public void arrayOfStringDeserialization() throws Exception {
-		final String className = "edu.cornell.mannlib.vitro.webapp.dynapi.data.types.DynapiArrayList";
+		final String className = DynapiArrayList.class.getCanonicalName();
 		ParameterType type = createArrayType("deserialize", "input type", true, "java.lang.String", "toString", "", false, false);
 		final String input = "[\"42\", \"42\"]";
 		Object result = ParameterConverter.deserialize(type, input);
@@ -90,7 +91,7 @@ public class ParameterConverterTest {
 
 	@Test
 	public void arrayOfStringSerialization() throws Exception {
-		final String className = "edu.cornell.mannlib.vitro.webapp.dynapi.data.types.DynapiArrayList";
+		final String className = DynapiArrayList.class.getCanonicalName();
 		ParameterType type = createArrayType("serialize", "input", true, "java.lang.String", "toString", "", false, true);
 		ArrayList input = new ArrayList(Arrays.asList("42", "42"));
 		Object result = ParameterConverter.serialize(type, input);
@@ -100,7 +101,7 @@ public class ParameterConverterTest {
 	
 	@Test
 	public void arrayOfIntegerDeserialization() throws Exception {
-		final String className = "edu.cornell.mannlib.vitro.webapp.dynapi.data.types.DynapiArrayList";
+		final String className = DynapiArrayList.class.getCanonicalName();
 		ParameterType type = createArrayType("deserialize", "input type", true, "java.lang.Integer", "parseInt", "input", true, false);
 		final String input = "[42, 42]";
 		Object result = ParameterConverter.deserialize(type, input);
@@ -110,7 +111,7 @@ public class ParameterConverterTest {
 
 	@Test
 	public void arrayOfIntegerSerialization() throws Exception {
-		final String className = "edu.cornell.mannlib.vitro.webapp.dynapi.data.types.DynapiArrayList";
+		final String className = DynapiArrayList.class.getCanonicalName();
 		ParameterType type = createArrayType("serialize", "input", true, "java.lang.Integer", "toString", "", false, true);
 		ArrayList input = new ArrayList(Arrays.asList(42, 42));
 		Object result = ParameterConverter.serialize(type, input);
@@ -120,7 +121,7 @@ public class ParameterConverterTest {
 
 	private ParameterType createArrayType(String arrayMethod, String arrayArgs, boolean arrayIsStatic,
 	String elementClassName, String elemMethodName, String elemArgs, boolean elemIsStatic, boolean serialization) throws Exception {
-		final String arrayClassName = "edu.cornell.mannlib.vitro.webapp.dynapi.data.types.DynapiArrayList";
+		final String arrayClassName = DynapiArrayList.class.getCanonicalName();
 		ArrayParameterType arrayType = new ArrayParameterType();
 		ImplementationType arrayImplType = new ImplementationType();
 		arrayType.setImplementationType(arrayImplType);
